refactor: adminboards 구조 변경
