+++ admin/src/main/java/com/kernel/admin/controller/board/AdminBoardsController.java
@@ -4,10 +4,10 @@ package com.kernel.admin.controller.board;
 import com.kernel.admin.domain.entity.Board;
 import com.kernel.admin.domain.enumerate.BoardType;
 import com.kernel.admin.repository.AdminBoardRepository;
-import com.kernel.admin.service.AdminBoardsService;
+import com.kernel.admin.service.board.AdminBoardsService;
 import com.kernel.admin.service.dto.request.AdminBoardsReqDTO;
 import com.kernel.admin.service.dto.responsse.AdminBoardsResDTO;
-import com.kernel.common.global.entity.ApiResponse;
+import com.kernel.global.service.dto.responsse.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
