+++ admin/src/main/java/com/kernel/admin/service/board/AdminBoardsService.java
@@ -1,4 +1,4 @@
-package com.kernel.admin.service;
+package com.kernel.admin.service.board;
 
 import com.kernel.admin.domain.enumerate.BoardType;
 import com.kernel.admin.service.dto.request.AdminBoardsReqDTO;
